Add channels block height open heuristic
