+++ agent/local/channel.go
@@ -2,11 +2,10 @@ package local
 
 import (
 	"context"
-	"encoding/binary"
-	"math/big"
 	"time"
 
 	"github.com/aftermath2/hydrus/config"
+	"github.com/aftermath2/hydrus/graph"
 	"github.com/aftermath2/hydrus/lightning"
 
 	"github.com/lightningnetwork/lnd/lnrpc"
@@ -68,7 +67,7 @@ func getChannels(
 			ID:              channel.ChanId,
 			Point:           channel.ChannelPoint,
 			Active:          channel.Active,
-			Age:             GetChannelBlockHeight(channel.ChanId),
+			Age:             graph.GetChannelBlockHeight(channel.ChanId),
 			Capacity:        uint64(channel.Capacity),
 			NumForwards:     numForwards,
 			ForwardsAmount:  forwardsAmount,
@@ -146,13 +145,4 @@ func getPeerInfo(channel *lnrpc.Channel, peers []*lnrpc.Peer) (int64, int32) {
 	}
 
 	return pingTime, flapCount
-}
-
-// GetChannelBlockHeight returns the block height at which a channel has been established based on its ID.
-func GetChannelBlockHeight(channelID uint64) uint32 {
-	idBytes := make([]byte, 8)
-	binary.BigEndian.PutUint64(idBytes, channelID)
-
-	blockHeight := new(big.Int).SetBytes(idBytes[0:3]).Uint64()
-	return uint32(blockHeight)
 }
