+++ agent/local/channel_test.go
@@ -403,40 +403,3 @@ func TestGetPeerInfo(t *testing.T) {
 		})
 	}
 }
-
-func TestGetChannelBlockHeight(t *testing.T) {
-	tests := []struct {
-		name      string
-		channelID uint64
-		expected  uint32
-		note      string
-	}{
-		{
-			name:      "Minimum value",
-			channelID: 0,
-			expected:  0,
-		},
-		{
-			name:      "Maximum uint64",
-			channelID: (1 << 64) - 1,
-			expected:  0xFFFFFF,
-		},
-		{
-			name:      "Regular channel ID",
-			channelID: 191315023298560,
-			expected:  174,
-		},
-		{
-			name:      "Hex channel ID",
-			channelID: 0x0001110000000000,
-			expected:  0x111,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			actual := GetChannelBlockHeight(tt.channelID)
-			assert.Equal(t, tt.expected, actual)
-		})
-	}
-}
