+++ docker/mainnet/hydrus.yml
@@ -37,6 +37,7 @@ agent:
       inbound_fee_rate: 0.8
       min_htlc: 1
       max_htlc: 0.5
+      block_height: 1
     close:
       capacity: 0.9
       active: 1
