+++ graph/graph_test.go
@@ -72,7 +72,7 @@ func TestNew(t *testing.T) {
 		},
 		Heuristics: graph.Heuristics{
 			Capacity:              heuristic.NewFull[uint64](1_000_000, 1_000_000, config.DefaultOpenWeights.Capacity, false),
-			Features:              heuristic.NewFull[int](0, 1, config.DefaultOpenWeights.Features, false),
+			Features:              heuristic.NewFull(0, 1, config.DefaultOpenWeights.Features, false),
 			Hybrid:                heuristic.NewFull(0, 1, config.DefaultOpenWeights.Hybrid, false),
 			BaseFee:               heuristic.NewFull[uint64](0, 1000, config.DefaultOpenWeights.BaseFee, true),
 			FeeRate:               heuristic.NewFull[uint64](100, 300, config.DefaultOpenWeights.FeeRate, true),
@@ -80,6 +80,7 @@ func TestNew(t *testing.T) {
 			InboundFeeRate:        heuristic.NewFull[int64](0, 10000, config.DefaultOpenWeights.InboundFeeRate, true),
 			MinHTLC:               heuristic.NewFull[uint64](1, 1, config.DefaultOpenWeights.MinHTLC, true),
 			MaxHTLC:               heuristic.NewFull[uint64](500_000, 1_000_000, config.DefaultOpenWeights.MaxHTLC, false),
+			BlockHeight:           heuristic.NewFull[uint64](0, 0, config.DefaultOpenWeights.BlockHeight, true),
 			DegreeCentrality:      heuristic.NewFull[float64](1, 1, config.DefaultOpenWeights.DegreeCentrality, false),
 			BetweennessCentrality: heuristic.NewFull[float64](0, 0, config.DefaultOpenWeights.BetweennessCentrality, false),
 			EigenvectorCentrality: heuristic.NewFull[uint64](1, 1, config.DefaultOpenWeights.EigenvectorCentrality, false),
@@ -280,6 +281,43 @@ func TestGetAddresses(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			got := graph.GetAddresses(tt.input)
 			assert.Equal(t, tt.expected, got)
+		})
+	}
+}
+
+func TestGetChannelBlockHeight(t *testing.T) {
+	tests := []struct {
+		name      string
+		channelID uint64
+		expected  uint32
+		note      string
+	}{
+		{
+			name:      "Minimum value",
+			channelID: 0,
+			expected:  0,
+		},
+		{
+			name:      "Maximum uint64",
+			channelID: (1 << 64) - 1,
+			expected:  0xFFFFFF,
+		},
+		{
+			name:      "Regular channel ID",
+			channelID: 191315023298560,
+			expected:  174,
+		},
+		{
+			name:      "Hex channel ID",
+			channelID: 0x0001110000000000,
+			expected:  0x111,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			actual := graph.GetChannelBlockHeight(tt.channelID)
+			assert.Equal(t, tt.expected, actual)
 		})
 	}
 }
